AU-2354: feature: AU-2354: Moved createFormData and parseMultiValueKey to form_data_helpers from form_helpers.ts.
